requirements.txt
insightface
onnxruntime
opencv-python
numpy
openai-whisper
google-genai
Pillow
rich
python-dotenv
